backend/requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
pydantic==2.9.2
sentence-transformers==3.2.0
pinecone-client==5.0.1
python-dotenv==1.0.1
google-cloud-translate==3.16.0